requirements.txt
Pillow>=10.0
rembg>=2.0
onnxruntime>=1.17
python-magic>=0.4
pydantic-settings>=2.0
structlog>=24.0
httpx>=0.27

# Dev / testing
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=5.0
ruff>=0.4
